# static/swagger.json
{"info": {"description": "A simple API for posts", "title": "Posts API", "version": "1.0.0"}, "paths": {"/posts": {"post": {"summary": "Create new post", "description": "Create new post", "tags": ["Posts"], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PostRequestBody"}}}}, "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Post"}}}}, "201": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Post"}}}}}}, "get": {"summary": "Get all posts", "description": "Get all posts", "tags": ["Posts"], "responses": {"200": {"content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Post"}}}}}}}}, "/posts/{post_id}": {"get": {"summary": "Get post by post id", "description": "Get post by post id", "tags": ["Posts"], "parameters": [{"in": "path", "name": "post_id", "schema": {"type": "string"}, "required": true}], "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Post"}}}}, "404": {"content": {"application/json": {"schema": {"message": "string"}}}}}}, "put": {"summary": "Update post with post id", "description": "Update post with post id", "tags": ["Posts"], "parameters": [{"in": "path", "name": "post_id", "schema": {"type": "string"}, "required": true}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PostRequestBody"}}}}, "responses": {"200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Post"}}}}, "404": {"content": {"application/json": {"schema": {"message": "string"}}}}}}, "delete": {"summary": "Delete post with post id", "description": "Delete post with post id", "tags": ["Posts"], "parameters": [{"in": "path", "name": "post_id", "schema": {"type": "string"}, "required": true}], "responses": {"200": {"content": {"application/json": {"schema": {"message": "string"}}}}, "404": {"content": {"application/json": {"schema": {"message": "string"}}}}}}}}, "openapi": "3.0.0", "components": {"schemas": {"PostRequestBody": {"type": "object", "properties": {"content": {"type": "string"}, "title": {"type": "string"}}}, "Post": {"type": "object", "properties": {"content": {"type": "string"}, "id": {"type": "integer"}, "title": {"type": "string"}}}}}}